requirements.txt
garminconnect>=0.2.0
PyYAML>=6.0
APScheduler>=3.10.0
openpyxl>=3.1.0
